+++ internal/entities/feed/entry.go
