+++ internal/storage/manticore/manticore.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 	openapiclient "github.com/manticoresoftware/manticoresearch-go"
-	"github.com/terratensor/feed-parser/service/internal/entities/feed"
+	"github.com/terratensor/feed-parser/internal/entities/feed"
 
 	"log"
 	"os"
