+++ internal/workerpool/task.go
@@ -3,10 +3,10 @@ package workerpool
 import (
 	"context"
 	"fmt"
-	"github.com/terratensor/feed-parser/service/internal/crawler"
-	"github.com/terratensor/feed-parser/service/internal/entities/feed"
-	"github.com/terratensor/feed-parser/service/internal/lib/logger/sl"
-	"github.com/terratensor/feed-parser/service/internal/storage/manticore"
+	"github.com/terratensor/feed-parser/internal/crawler"
+	"github.com/terratensor/feed-parser/internal/entities/feed"
+	"github.com/terratensor/feed-parser/internal/lib/logger/sl"
+	"github.com/terratensor/feed-parser/internal/storage/manticore"
 	"log"
 	"log/slog"
 	"os"
